Add default message handler that can be used by both the key beep mixin and switch2477s at the same time. Use absolute imports because that seems to be what makes jython happy.
