+++ python/switch2477S.py
@@ -9,22 +9,13 @@ from querier import Querier
 from querier import MsgHandler
 from switch import Switch
 from linkdb import *
-from mixins.key_beep_mixin import KeyBeepMixin
+
+from python.handlers.defaultmessagehandler import DefaultMsgHandler
+from python.mixins.keybeepmixin import KeyBeepMixin
 
 from us.pfrommer.insteon.msg import Msg
 from us.pfrommer.insteon.msg import MsgListener
 from us.pfrommer.insteon.msg import InsteonAddress
-
-def out(msg = ""):
-	iofun.out(msg)
-
-class DefaultMsgHandler(MsgHandler):
-	label = None
-	def __init__(self, l):
-		self.label = l
-	def processMsg(self, msg):
-		out(self.label + " got msg: " + msg.toString())
-		return 1
 
 class Switch2477S(Switch, KeyBeepMixin):
 	"""==============  Insteon SwitchLinc 2477S ==============="""
